+++ src/screens/MediaSettings/Components/EPub/ChapterNavigator/index.scss
@@ -1,0 +1,88 @@
+////////////////////////////////////////////////////////////
+// Chapter Navigator
+////////////////////////////////////////////////////////////
+.msp-ee-cn-con {
+  position: fixed;
+  top: 150px;
+  left: 0;
+  z-index: 3;
+
+  .ee-cn-open-btn {
+    animation: ee-cn-show 300ms linear;
+    &:hover span {
+      box-shadow: 0 0 6px rgba(0, 0, 0, 0.459);
+    }
+
+    span {
+      border-top-left-radius: 0 !important;
+      border-bottom-left-radius: 0 !important;
+      box-shadow: 0 0 4px rgba(0, 0, 0, 0.384);
+      // background-color: rgb(38, 148, 148) !important;
+      
+      i {
+        font-size: 30px !important;
+      }
+    }
+  }
+
+  .ee-cn-wrapper {
+    position: fixed;
+    width: 100vw;
+    height: 100vh;
+    top: 0;
+    left: 0;
+  }
+
+  .ee-cn-ch-con {
+    width: 270px;
+    padding: 0 20px 20px 0;
+    margin-top: -30px;
+    // background-color: rgb(226, 226, 226);
+    background-color: rgb(242, 246, 246);
+    border-radius: 0 20px 20px 0;
+    box-shadow: 0 0 15px rgba(0, 0, 0, 0.295);
+    height: 600px;
+    max-height: 70vh;
+    animation: ee-cn-show 100ms linear;
+    overflow-x: hidden;
+    overflow-y: auto;
+
+    &.hide {
+      margin-left: -270px;
+      animation: ee-cn-hide 100ms linear;
+    }
+
+    .ee-cn-h {
+      justify-content: space-between;
+      align-items: center;
+      background: linear-gradient(-180deg,rgb(242, 246, 246) 90%,transparent);
+      position: sticky;
+      top: 0;
+      left: 0;
+      z-index: 1;
+      padding: 10px 0 10px 20px;
+      h3 {
+        padding: 0;
+        margin: 0;
+      }
+    }
+
+    .ee-cn-ch-ul {
+      .ee-cn-ch-li span {
+        padding-left: 10px;
+        text-align: left;
+        border-top-left-radius: 0;
+        border-bottom-left-radius: 0;
+      }
+    }
+  }
+}
+
+@keyframes ee-cn-show {
+  0% { margin-left: -270px; }
+  100% { margin-left: 0px; }
+}
+@keyframes ee-cn-hide {
+  0% { margin-left: 0px; }
+  100% { margin-left: -270px; }
+}
