+++ src/screens/MediaSettings/Components/EPub/CoverPicker.jsx
