+++ src/screens/MediaSettings/Components/EPub/EpubChapters/index.js
@@ -1,12 +1,20 @@
 import React from 'react'
+import _ from 'lodash'
 import EpubList from './EpubList'
+import { Dropdown } from 'semantic-ui-react'
 import { Button } from 'pico-ui'
+import { langMap } from 'screens/Watch/Utils'
+import './index.scss'
+
+const langOptions = _.map(langMap, (text, value) => ({ text, value }))
 
 export default function EpubChapters({
   chapters=[],
   splitChapter,
   undoSplitChapter,
   changeChapter,
+  language,
+  changeLanguage,
   handleTitleChange,
   magnifyImage,
   endMagnifyImage,
@@ -16,10 +24,21 @@ export default function EpubChapters({
   unfoldChapter
 }) {
 
+  // console.log('chapters', chapters)
+
   return (
     <div className="msp-ee-el-con ct-d-c">
       <div className="msp-ee-el-header">
-        <h1>Manage Your ePub Book</h1>
+        <div className="w-100 ct-d-r-center-v msp-ee-el-h1">
+          <h1>Manage Your ePub Book</h1>
+          <Dropdown pointing='right'
+            trigger={<Button outlined icon="arrow_drop_down" text={langMap[language]} />}
+            icon={null}
+            value={language}
+            options={langOptions}
+            onChange={(e, { value }) => changeLanguage(value)}
+          />
+        </div>
         <p>
           <span className="msp-ee-el-h-p-t">Instruction</span>
           To manage your ePub chapters, set <span>splitting points</span> between screenshots to generate an initial version of ePub chapters.
