+++ src/screens/MediaSettings/Components/EPub/EpubChapters/index.scss
@@ -1,22 +1,3 @@
-.msp-ee-con {
-  width: 100%;
-  position: relative;
-  padding-top: 93px;
-  min-width: 1100px;
-  height: 100vh;
-  // background-color: rgba(248, 248, 248, 0.459);
-
-  .msp-ee {
-    overflow-y: auto;
-    overflow-x: hidden;
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    height: calc(100vh - 93px);
-  }
-}
-
-
 ////////////////////////////////////////////////////////////
 // Chapter List
 ////////////////////////////////////////////////////////////
@@ -54,12 +35,18 @@
   .msp-ee-el-header {
     margin: 0 0 30px 100px;
     padding-top: 30px;
-    h1 {
-      font-size: 40px;
+    .msp-ee-el-h1 {
+      align-items: center;
+      h1 {
+        font-size: 40px;
+        margin: 10px 20px 10px 0;
+      }
     }
+
     p {
       font-size: 18px;
       border-left: 4px solid rgba(31, 163, 163, 0.849);
+      margin-top: 1rem;
       padding: .5rem 1rem;
       background-color: rgba(24, 131, 131, 0.068);
       border-radius: 0 10px 10px 0;
@@ -79,7 +66,7 @@
   .ee-el-chapters {
     position: relative;
     width: 100%;
-    padding-top: 2rem;
+    padding-top: 1rem;
 
     .ee-el-items {
       position: relative;
@@ -229,205 +216,4 @@
       }
     }
   }
-}
-
-
-////////////////////////////////////////////////////////////
-// ePub Preview
-////////////////////////////////////////////////////////////
-.msp-ee-ep-con {
-  position: relative;
-  width: 45%;
-  padding: 60px 30px 150px 30px;
-  position: sticky;
-  top: 0;
-  height: calc(100vh - 93px);
-  overflow-y: auto;
-  overflow-x: visible;
-}
-
-
-////////////////////////////////////////////////////////////
-// Action Buttons
-////////////////////////////////////////////////////////////
-.msp-ee-act {
-  position: fixed;
-  right: 20px;
-  bottom: 15px;
-
-  .ee-act-btn span {
-    box-shadow: 0 0 15px rgba(0, 0, 0, 0.384);
-  }
-
-  .ee-act-save-btn {
-    margin-right: 10px;
-  }
-}
-
-////////////////////////////////////////////////////////////
-// Cover Picker
-////////////////////////////////////////////////////////////
-.msp-ee-cp-con {
-  width: 100%;
-  display: flex;
-
-  .ee-cp-imgs {
-    width: 60%;
-    display: -webkit-box;
-    display: -webkit-flex;
-    display: flex;
-    -webkit-flex-direction: row;
-    flex-direction: row;
-    -webkit-flex-wrap: wrap;
-    flex-wrap: wrap;
-    -webkit-justify-content: center;
-    justify-content: center;
-    align-items: flex-start;
-    max-height: 400px;
-    overflow-y: auto;
-    overflow-x: hidden;
-
-    .ee-cp-img-con {
-      position: relative;
-      margin: 5px;
-      cursor: pointer;
-
-      img {
-        width: 100px;
-      }
-
-      .ee-cp-img-wrapper {
-        position: absolute;
-        width: 100%;
-        height: 100%;
-        top: 0;
-        left: 0;
-
-        &:hover {
-          background-color: rgba(0, 0, 0, 0.287);
-        }
-
-        i {
-          font-size: 35px;
-          color: white;
-        }
-      }
-
-      &[data-current=true] .ee-cp-img-wrapper {
-        background-color: rgba(0, 0, 0, 0.61);
-      }
-    }
-  }
-
-  .ee-cp-cover {
-    width: 40%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-
-    img {
-      width: 100%;
-    }
-  }
-}
-
-////////////////////////////////////////////////////////////
-// Chapter Navigator
-////////////////////////////////////////////////////////////
-.msp-ee-cn-con {
-  position: fixed;
-  top: 150px;
-  left: 0;
-  z-index: 3;
-
-  .ee-cn-open-btn {
-    animation: ee-cn-show 300ms linear;
-    &:hover span {
-      box-shadow: 0 0 6px rgba(0, 0, 0, 0.459);
-    }
-
-    span {
-      border-top-left-radius: 0 !important;
-      border-bottom-left-radius: 0 !important;
-      box-shadow: 0 0 4px rgba(0, 0, 0, 0.384);
-      // background-color: rgb(38, 148, 148) !important;
-      
-      i {
-        font-size: 30px !important;
-      }
-    }
-  }
-
-  .ee-cn-wrapper {
-    position: fixed;
-    width: 100vw;
-    height: 100vh;
-    top: 0;
-    left: 0;
-  }
-
-  .ee-cn-ch-con {
-    width: 270px;
-    padding: 0 20px 20px 0;
-    margin-top: -30px;
-    // background-color: rgb(226, 226, 226);
-    background-color: rgb(242, 246, 246);
-    border-radius: 0 20px 20px 0;
-    box-shadow: 0 0 15px rgba(0, 0, 0, 0.295);
-    height: 600px;
-    max-height: 70vh;
-    animation: ee-cn-show 100ms linear;
-    overflow-x: hidden;
-    overflow-y: auto;
-
-    &.hide {
-      margin-left: -270px;
-      animation: ee-cn-hide 100ms linear;
-    }
-
-    .ee-cn-h {
-      justify-content: space-between;
-      align-items: center;
-      background: linear-gradient(-180deg,rgb(242, 246, 246) 90%,transparent);
-      position: sticky;
-      top: 0;
-      left: 0;
-      z-index: 1;
-      padding: 10px 0 10px 20px;
-      h3 {
-        padding: 0;
-        margin: 0;
-      }
-    }
-
-    .ee-cn-ch-ul {
-      .ee-cn-ch-li span {
-        padding-left: 10px;
-        text-align: left;
-        border-top-left-radius: 0;
-        border-bottom-left-radius: 0;
-      }
-    }
-  }
-}
-
-@keyframes ee-cn-show {
-  0% { margin-left: -270px; }
-  100% { margin-left: 0px; }
-}
-@keyframes ee-cn-hide {
-  0% { margin-left: 0px; }
-  100% { margin-left: -270px; }
-}
-@keyframes ee-el-ch-hide {
-  0% { padding: 200px 0; opacity: 0; }
-  100% { padding: 0px 0; opacity: 1; }
-}
-@keyframes ee-el-ch-show {
-  0% { opacity: 0; max-height: 50px; }
-  100% { opacity: 1; max-height: 1000px; }
-}
-@keyframes ee-el-title-show {
-  0% { padding: 0 0 0 90px; opacity: 0; }
-  100% { padding: 15px 0 15px 90px; opacity: 1; }
 }
